refactor(All): 게임과 추측으로 패키지 분리
