+++ src/main/java/leets/land/domain/guess/GuessAlphabet.java
@@ -1,23 +1,23 @@
-package leets.land.domain;
+package leets.land.domain.guess;
 
 import java.util.Objects;
 
-public class GuessNumber implements GuessValue {
+public class GuessAlphabet implements GuessValue {
 
-    private static final int MIN_RANGE = 1;
-    private static final int MAX_RANGE = 100;
+    private static final int MIN_RANGE = 'A';
+    private static final int MAX_RANGE = 'z';
     private static final String ERROR_FORMAT = "[ERROR] %s";
 
     private final int value;
 
-    public GuessNumber(int value) {
+    public GuessAlphabet(int value) {
         validate(value);
         this.value = value;
     }
 
     private void validate(int value) {
         if (value < MIN_RANGE || value > MAX_RANGE) {
-            throw new IllegalArgumentException(String.format(ERROR_FORMAT, "추측값의 범위는 1 이상 100 이하입니다"));
+            throw new IllegalArgumentException(String.format(ERROR_FORMAT, "추측값의 범위는 A 이상 z 이하입니다"));
         }
     }
 
@@ -33,7 +33,7 @@ public class GuessNumber implements GuessValue {
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
-        GuessNumber that = (GuessNumber) o;
+        GuessAlphabet that = (GuessAlphabet) o;
         return value == that.value;
     }
