+++ src/main/java/leets/land/domain/updowngame/NumberUpDownGame.java
@@ -1,5 +1,8 @@
-package leets.land.domain;
+package leets.land.domain.updowngame;
 
+import leets.land.domain.AnswerNumber;
+import leets.land.domain.guess.GuessNumber;
+import leets.land.domain.guess.GuessRange;
 import leets.land.domain.status.GuessStatus;
 
 public class NumberUpDownGame implements UpDownGame {
