+++ src/main/java/leets/land/domain/updowngame/UpDownGame.java
@@ -1,5 +1,7 @@
-package leets.land.domain;
+package leets.land.domain.updowngame;
 
+import leets.land.domain.guess.GuessNumber;
+import leets.land.domain.guess.GuessRange;
 import leets.land.domain.status.GuessStatus;
 
 public interface UpDownGame {
