+++ src/test/java/leets/land/domain/AnswerNumberTest.java
@@ -1,5 +1,6 @@
 package leets.land.domain;
 
+import leets.land.domain.guess.GuessNumber;
 import leets.land.domain.status.GuessStatus;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
