+++ src/test/java/leets/land/domain/GuessNumberTest.java
@@ -1,5 +1,6 @@
 package leets.land.domain;
 
+import leets.land.domain.guess.GuessNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
